Eliminate Type::Coercion
Type coercion is extremely fragile, as it does not support recursive
coercion or properly solves alternatives (e.g. Dict[key => coercible] |
Dict [key2 => coercible]). Up to now, it was only used to convert
strings to Path::Tiny objects. We do this manually now in the BUILD
steps of the affected modules.
Sometimes, life is more pleasant with less magic.