Igor::Config: Bugfix: handle relative paths correctly
Relative paths in configurations files are supposed to be relative
w.r.t. the config file containing the path, e.g. the main configuration
file for repositories, or the package.toml files for packages within the
repo. This commit implements correct handling for these cases.
This includes only symlinking absolute and not relative paths.