Add vault concept for secret-management
This commit adds a basic vault concept to igor. Vaults can be used to
hide passwords within encrypted files, which are only decrypted on
demand and can then be interpolated into templated files.
Vaults can be cached for ease of use. Caching is based on the checksum
of the original, encrypted files. A new decryption will be required
when the vault file changes.